Updated driver from pigpio to libgpiod

+++ src/VL6180xcallbackChild.cpp
@@ -1,6 +1,6 @@
 #include <stdio.h>
 #include <VL6180xcallbackChild.h>
 
-void VL6180xcallbackChild::hasSample(uint8_t v){
+void VL6180xcallbackChild::hasSample(int v){
 	printf("Distance to object is %u mm. \n", v);  //Changing this had no effect on Segmentation Fault error.
 };
